Some modification suggested by code maze team.
